+++ package.json
@@ -14,6 +14,7 @@
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react-native-collapsible": "^1.5.1",
     "react-native-elements": "^1.2.6",
     "react-native-gesture-handler": "~1.3.0",
     "react-native-reanimated": "~1.2.0",
